made more in line with my repo
