+++ control/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "control"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1.0.198", features = ["derive"] }
+logging = { path = "../logging" }
