+++ logging/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "logging"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = "1.0.198"
