+++ logging/src/lib.rs
@@ -1,18 +1,8 @@
-use serde::Serialize;
+use serde::ser::Serialize;
 use std::fmt::Debug;
 use std::sync::Arc;
 
-#[derive(Serialize, Debug, Clone)]
-struct ControlState {
-    pub time: f64,
-    pub update_count: u32,
-}
-
-struct CascadedControl {
-    logger: Arc<StateLogger<ControlState>>,
-}
-
-struct StateLogger<T> {
+pub struct StateLogger<T> {
     name: String,
     buffer_size: usize,
     _marker: std::marker::PhantomData<T>,
@@ -28,7 +18,7 @@ impl<T> StateLogger<T> {
     }
 }
 
-struct RustStateLoggerManager;
+pub struct RustStateLoggerManager;
 
 impl RustStateLoggerManager {
     pub fn create_logger<T: Serialize + Debug + Send + Sync + Clone + 'static>(
@@ -38,17 +28,4 @@ impl RustStateLoggerManager {
     ) -> Arc<StateLogger<T>> {
         Arc::new(StateLogger::new(name.to_string(), buffer_size))
     }
-}
-
-impl CascadedControl {
-    pub fn new(logger_manager: &RustStateLoggerManager) -> Self {
-        let logger = logger_manager.create_logger::<ControlState>("control/cascaded_control", 1000);
-
-        CascadedControl { logger }
-    }
-}
-
-fn main() {
-    let logger_manager = RustStateLoggerManager;
-    let control = CascadedControl::new(&logger_manager);
 }
